+++ modules/qotd/misc.ts
